Addressed Pull Request comments
